agent: avoid infinite loop when poll()ing
We should always return some value from each iteration of poll().

If the current time was equivalent to the send time at poll(), it was
possible to infinite loop.

If retransmissions of a transaction were cancelled, we were also not
waiting the full length of the timeout for a response.